Partially fixed #5824 - "Add function for unloading and reloading San Andreas map" Added: removeWorldBuilding ( modelID, radius, x, y, z ) restoreWorldBuilding ( modelID, radius, x, y, z ) restoreWorldBuildings ( )
All client side 

Note: Only works on Binary IPL objects for now ( Binary IPL objects make up the vast majority of the map and a full list can be found here: http://www.ukscifi.net/A/gta/GTADEV/Scripts/binipl2.txt ) - Does not apply if you are using r3568 or greater.
Note 2: To remove a building call removeWorldBuilding on it when it is streamed out else it will need to be streamed out then back in.
Note 3: Try to keep radius small as possible.
Note 4: The names are still in debate as I've currently just come out of an hour long conversation about the function names but it's 3am and I need some sleep.

Merry Christmas and HF.

- Caz